+++ src/crystal/progress/ui.py
@@ -1,6 +1,6 @@
 from collections.abc import Callable
+from crystal.progress.interface import CancelLoadUrls, CancelOpenProject, CancelSaveAs, LoadUrlsProgressListener, OpenProjectProgressListener, SaveAsProgressListener, VetoUpgradeProject
 from crystal.util.bulkheads import capture_crashes_to_stderr
-from crystal.util.headless import is_headless_mode
 from crystal.util.sizes import format_byte_size
 from crystal.util.wx_dialog import position_dialog_initially, ShowModal
 from crystal.util.xfunctools import partial2
@@ -135,61 +135,7 @@ class _AbstractProgressDialog:
 
 
 # ------------------------------------------------------------------------------
-# OpenProjectProgressListener
-
-_active_progress_listener = None  # type: Optional[OpenProjectProgressListener]
-
-
-class CancelOpenProject(Exception):
-    pass
-
-
-class VetoUpgradeProject(Exception):
-    pass
-
-
-# NOTE: See subclass OpenProjectProgressDialog for the documentation of
-#       the various methods in this interface.
-class OpenProjectProgressListener:
-    def opening_project(self) -> None:
-        pass
-    
-    def upgrading_project(self, message: str) -> None:
-        pass
-    
-    def will_upgrade_revisions(self, approx_revision_count: int, can_veto: bool) -> None:
-        pass
-    
-    def upgrading_revision(self, index: int, revisions_per_second: float) -> None:
-        pass
-    
-    def upgrade_revisions_disk_error(self) -> None:
-        pass
-    
-    def did_upgrade_revisions(self, revision_count: int) -> None:
-        pass
-    
-    def loading_root_resources(self, root_resource_count: int) -> None:
-        pass
-    
-    def loading_resource_groups(self, resource_group_count: int) -> None:
-        pass
-    
-    def loading_root_resource_views(self) -> None:
-        pass
-    
-    def loading_resource_group_views(self) -> None:
-        pass
-    
-    def creating_entity_tree_nodes(self, entity_tree_node_count: int) -> None:
-        pass
-    
-    def reset(self) -> None:
-        pass
-
-
-DummyOpenProjectProgressListener = OpenProjectProgressListener
-
+# OpenProjectProgressDialog
 
 class OpenProjectProgressDialog(_AbstractProgressDialog, OpenProjectProgressListener):
     _dialog_title = 'Opening Project...'  # override
@@ -442,33 +388,7 @@ class OpenProjectProgressDialog(_AbstractProgressDialog, OpenProjectProgressList
 
 
 # ------------------------------------------------------------------------------
-# LoadUrlsProgressListener
-
-class CancelLoadUrls(Exception):
-    pass
-
-
-# NOTE: See subclass LoadUrlsProgressDialog for the documentation of
-#       the various methods in this interface.
-class LoadUrlsProgressListener:
-    def will_load_resources(self, approx_resource_count: int) -> None:
-        pass
-    
-    def loading_resource(self, index: int) -> None:
-        pass
-    
-    def did_load_resources(self, resource_count: int) -> None:
-        pass
-    
-    def indexing_resources(self) -> None:
-        pass
-    
-    def reset(self) -> None:
-        pass
-
-
-DummyLoadUrlsProgressListener = LoadUrlsProgressListener
-
+# LoadUrlsProgressDialog
 
 class LoadUrlsProgressDialog(_AbstractProgressDialog, LoadUrlsProgressListener):
     _dialog_title = 'Loading URLs...'  # override
@@ -524,34 +444,7 @@ class LoadUrlsProgressDialog(_AbstractProgressDialog, LoadUrlsProgressListener):
 
 
 # ------------------------------------------------------------------------------
-# SaveAsProgressListener
-
-class CancelSaveAs(Exception):
-    """Raised when a save operation is canceled by the user."""
-    pass
-
-
-# NOTE: See subclass SaveAsProgressDialog for the documentation of
-#       the various methods in this interface.
-class SaveAsProgressListener:
-    def calculating_total_size(self, message: str) -> None:
-        pass
-    
-    def total_size_calculated(self, total_file_count: int, total_byte_count: int) -> None:
-        pass
-    
-    def copying(self, file_index: int, filename: str, bytes_copied: int, bytes_per_second: float) -> None:
-        pass
-    
-    def did_copy_files(self) -> None:
-        pass
-    
-    def reset(self) -> None:
-        pass
-
-
-DummySaveAsProgressListener = SaveAsProgressListener
-
+# SaveAsProgressDialog
 
 def _cancel_or_fg_trampoline(func: Callable[..., None]) -> Callable[..., None]:
     """
@@ -775,10 +668,6 @@ class DeferredProgressDialog:
         be shown as a window modal dialog.
         """
         if self._shown:
-            return
-        
-        # Silently refuse to show if running in headless mode
-        if is_headless_mode():
             return
         
         # Show now. Apply deferred operations.